Rename few files, did some tests with rakefile.
Web server have to be implemented into cinch plugin.
