+++ Gemfile
@@ -8,3 +8,4 @@ gem 'dotenv'
 gem 'sinatra'
 gem 'json', '1.8.1'
 gem 'rack-test'
+gem 'bson_ext'
