+++ rakefile
@@ -1,5 +1,25 @@
+require 'rubygems'
+require 'sinatra'
+
 begin
   require 'rspec/core/rake_task'
   RSpec::Core::RakeTask.new(:spec)
 rescue LoadError
 end
+
+namespace :revuebot do
+  desc "Start Revue irc bot"
+  task :start do
+    system "ruby revuebot.rb"
+  end
+end
+
+namespace :sinatra do
+  desc "Start Sinatra web server"
+  task :start do
+    #run Sinatra::Application
+  end
+end
+
+desc "Start bot and webserver."
+multitask :start => [ 'sinatra:start', 'revuebot:start' ]
