+++ spec/spec_helper.rb
@@ -1,2 +1,2 @@
 require './lib/cinch/plugins/revuelog'
-require './testsinatra'
+require './webserver'
